Open up multirate filters even further
This is an attempt to produce a more natural sound by allowing a wider range of
input frequencies. I made this change a while ago, but never committed it, and
I don't recall how I felt about it at the time. I do know that I determined
that this hardware isn't powerful enough to achieve my target performance.